requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
argon2-cffi=20.1.0=py39h07f9747_2
async_generator=1.10=py_0
attrs=20.2.0=pyh9f0ad1d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
bleach=3.2.1=pyh9f0ad1d_0
brotlipy=0.7.0=py39h38d8fee_1001
bzip2=1.0.8=h516909a_3
c-ares=1.16.1=h516909a_3
ca-certificates=2020.12.5=ha878542_0
cairo=1.16.0=h9f066cc_1006
certifi=2020.12.5=py39hf3d152e_1
cffi=1.14.3=py39h0da4893_1
chardet=3.0.4=py39h079e4ff_1008
click=7.1.2=pyh9f0ad1d_0
cryptography=3.2.1=py39h01d1d65_0
curl=7.71.1=he644dc0_8
dbus=1.13.6=hfdff14a_1
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py39hde42818_1002
expat=2.2.9=he1b5a44_2
ffmpeg=4.3.1=h3215721_1
flask=1.1.2=pyh9f0ad1d_0
flask-cors=3.0.8=py_0
fontconfig=2.13.1=h7e3eb15_1002
freetype=2.10.4=h7ca028e_0
gettext=0.19.8.1=hf34092f_1004
giflib=5.2.1=h36c2ea0_2
glib=2.66.2=h58526e2_0
gmp=6.2.0=h58526e2_4
gnutls=3.6.13=h79a8f9a_0
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
harfbuzz=2.7.2=hb1ce69c_1
hdf5=1.10.6=nompi_h1022a3e_1110
icu=67.1=he1b5a44_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
ipykernel=5.3.4=py39h9cd51e4_1
ipython=7.19.0=py39hef51801_0
ipython_genutils=0.2.0=py_1
itsdangerous=1.1.0=py_0
jansson=2.11=h516909a_1001
jasper=1.900.1=h07fcdf6_1006
jedi=0.17.2=py39hf3d152e_1
jinja2=2.11.2=pyh9f0ad1d_0
jpeg=9d=h36c2ea0_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py39hde42818_2
jupyterlab=2.2.9=py_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=1.2.0=py_0
krb5=1.17.1=hfafb76e_3
lame=3.100=h14c3975_1001
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.35=h769bd43_9
leptonica=1.78.0=ha0c4403_3
libarchive=3.3.3=hddc7a2b_1008
libblas=3.9.0=2_openblas
libcblas=3.9.0=2_openblas
libclang=10.0.1=default_hde54327_1
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=hcdb4288_3
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=9.3.0=he4bcb1c_17
libgfortran5=9.3.0=he4bcb1c_17
libglib=2.66.2=hbe7bbb4_0
libgomp=9.3.0=h5dbcf3e_17
libiconv=1.16=h516909a_0
liblapack=3.9.0=2_openblas
liblapacke=3.9.0=2_openblas
libllvm10=10.0.1=he513fc3_3
libnghttp2=1.41.0=h8cfc5f6_2
libopenblas=0.3.12=pthreads_h4812303_1
libopencv=4.5.0=py39_2
libpng=1.6.37=h21135ba_2
libpq=12.3=h5513abc_2
libsodium=1.0.18=h516909a_1
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtiff=4.1.0=h4f3a223_6
libuuid=2.32.1=h14c3975_1000
libwebp=1.1.0=h56121f0_4
libwebp-base=1.1.0=h36c2ea0_3
libxcb=1.13=h14c3975_1002
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=h68273f3_2
lz4-c=1.9.2=he1b5a44_3
lzo=2.10=h516909a_1000
markupsafe=1.1.1=py39h38d8fee_2
mistune=0.8.4=py39h07f9747_1002
more-itertools=8.6.0=pypi_0
mysql-common=8.0.21=2
mysql-libs=8.0.21=hf3661c5_2
nbclient=0.5.1=py_0
nbconvert=6.0.7=py39hde42818_2
nbformat=5.0.8=py_0
ncurses=6.2=h58526e2_3
nest-asyncio=1.4.2=pyhd8ed1ab_0
nettle=3.4.1=h1bed415_1002
notebook=6.1.4=py39hde42818_1
nspr=4.29=he1b5a44_1
nss=3.58=h27285de_1
numpy=1.19.4=py39h57d35e7_1
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.0=py39_2
openh264=2.1.1=h8b12597_0
openjpeg=2.3.1=h981e76c_3
openssl=1.1.1i=h7f98852_0
packaging=20.4=pyh9f0ad1d_0
pandoc=2.11.0.4=hd18ef5c_0
pandocfilters=1.4.2=py_1
parso=0.7.1=pyh9f0ad1d_0
pcre=8.44=he1b5a44_0
pdf2image=1.13.1=py_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.0.1=py39h397ff4d_0
pip=20.2.4=py_0
pixman=0.40.0=h36c2ea0_0
poppler=20.12.1=h8330bde_1
poppler-data=0.4.10=0
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt-toolkit=3.0.8=pyha770c72_0
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py_1001
py-opencv=4.5.0=py39h9cd51e4_2
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.2=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py39h07f9747_1
pysocks=1.7.1=py39h079e4ff_2
pytesseract=0.3.6=pyh9f0ad1d_1
python=3.9.0=h2a148a8_4_cpython
python-dateutil=2.8.1=py_0
python_abi=3.9=1_cp39
pyzmq=19.0.2=py39hb69f2a1_2
qt=5.12.9=h1f2b2cb_0
readline=8.0=he28a2e2_2
requests=2.24.0=pyh9f0ad1d_0
scipy=1.5.3=py39hee8e79c_0
send2trash=1.5.0=py_0
setuptools=49.6.0=py39h079e4ff_2
six=1.15.0=pyh9f0ad1d_0
sqlite=3.33.0=h4cf870e_1
terminado=0.9.1=py39hde42818_1
tesseract=4.1.1=h58164bb_3
testpath=0.4.4=py_0
tk=8.6.10=hed695b0_1
tornado=6.1=py39hbd71b63_0
traitlets=5.0.5=py_0
tzdata=2020d=h516909a_0
urllib3=1.25.11=py_0
uwsgi=2.0.19.1=py39h257e732_1
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.35.1=pyh9f0ad1d_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.12=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
zeromq=4.3.3=he1b5a44_2
zipp=3.4.0=py_0
zlib=1.2.11=h516909a_1010
zstd=1.4.5=h6597ccf_2
